+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 - Add policies to allow dynamic scaling
 
+- Changed: (Breaking change) The default ECS image is changed to the latest amazon linux2 ami available at the time the terraform script is running. To ensure the version is fixed, provide the filter variable with to select an ami. See the [readme][README.md] for more details
+
+## 1.3.0
+- Update default base images to 2018.03.e
+
 ## 1.2.0
 - Add a second, more advanced example.
 
@@ -28,6 +33,7 @@ https://github.com/philips-software/terraform-aws-ecs/tags/1.0.0
 - Renamed output variables
 - Update of ECS optimized AMI's
 
-[Unreleased]: https://github.com/philips-software/terraform-aws-ecs/compare/1.2.0...HEAD
+[Unreleased]: https://github.com/philips-software/terraform-aws-ecs/compare/1.3.0...HEAD
+[1.3.0] https://github.com/philips-software/terraform-aws-ecscompare/1.2.0...1.3.0
 [1.2.0] https://github.com/philips-software/terraform-aws-ecscompare/1.1.0...1.2.0
 [1.1.0] https://github.com/philips-software/terraform-aws-ecscompare/1.0.0...1.1.0